+++ algorithms/models/ReportTypes.py
@@ -16,10 +16,34 @@ report_types: [Model] = [
         vector="attendance"
     ),
     Model(
+        key="attendances",
+        type="value",
+        text="presence",
+        vector="presence"
+    ),
+    Model(
         key="leaves",
         type="value",
         text="leave",
         vector="leave"
+    ),
+    Model(
+        key="leaves",
+        type="value",
+        text="vacation",
+        vector="vacation"
+    ),
+    Model(
+        key="leaves",
+        type="value",
+        text="permission",
+        vector="permission"
+    ),
+    Model(
+        key="leaves",
+        type="value",
+        text="permit",
+        vector="permit"
     ),
     Model(
         key="overtimes",