+++ algorithms/sbert.py
@@ -4,6 +4,7 @@ os.environ["KMP_DUPLICATE_LIB_OK"] = 'TRUE'
 import warnings
 from dataclasses import replace
 from spellchecker import SpellChecker
+from googletrans import Translator
 from sentence_transformers import SentenceTransformer
 import numpy as np
 import faiss
@@ -37,6 +38,7 @@ report_companies = [replace(re, vector=text_to_vector(re.vector).flatten()) for 
 
 # initialize requirement libraries
 spell = SpellChecker()
+translator = Translator()
 
 # perform search
 def run_algorithm(data, word_vector, total_results, similarity_scores):
@@ -94,36 +96,78 @@ def extraction(words: str):
     for i in range(len(words)):
         word = words[i]
         typo_possibility = False
+        diff_lang_possibility = False
 
         while True:
             if typo_possibility:
-                word = str(spell.candidates(word)) # check for the typo
+                word = str(spell.candidates(word)) # check for the typo (now only support for english words)
 
+            # perform to check each words
             res_report_types, res_report_employees, res_report_companies = check_word(word, res_report_types, res_report_employees, res_report_companies)
             if len(res_report_types) > 0 or len(res_report_employees) > 0 or len(res_report_companies) > 0:
                 break
             else:
-                if typo_possibility:
+                # loop steps
+                # 1. maybe language is different, so translate it
+                # 2. check typo in original word:lang
+                if diff_lang_possibility is False:
+                    translated = translator.translate(words[i], dest="en") # use `words[i]` because I want to translate original word
+                    if translated is not None:
+                        word = translated.text
+
+                    diff_lang_possibility = True
+                elif diff_lang_possibility and typo_possibility is False:
+                    word = words[i] # change word with original word (before translated)
+                    typo_possibility = True
+                elif typo_possibility and diff_lang_possibility:
                     break
 
-                typo_possibility = True
+    # response
+    return res_report_types, res_report_employees, res_report_companies
 
 
-    # log response
-    print("\nReport types possibility: ", res_report_types)
-    print("Employees possibility: ", res_report_employees)
-    print("Companies possibility: ", res_report_companies)
+def init(words: str, debug=True):
+    if debug:
+        print("user inputted: ", user_input)
+
+    # run algorithm for main data
+    res_report_types, res_report_employees, res_report_companies = extraction(user_input)
+
+    if debug:
+        # print response for core data
+        print("\nReport types possibility: ", res_report_types)
+        print("Employees possibility: ", res_report_employees)
+        print("Companies possibility: ", res_report_companies)
+
+    # run algorithm for date extraction
+    user_input_translated = translator.translate(user_input).text.replace("'s","") # remove possesive pronoun
+    date_extracted = extract_dates_and_range(user_input_translated)
+
+    if debug:
+        # log response for date extracted
+        print("user input translatted: ", user_input_translated)
+        print("Date parameter: ", date_extracted)
+
+    if debug is False:
+        # return as data for service purpose
+        return {
+            "cores": {
+                "report_types": res_report_types,
+                "report_employees": res_report_employees,
+                "report_companies": res_report_companies,
+            },
+            "date_extraction": date_extracted
+        }
 
 
-# get user input
-while True:
-    user_input: str = input("\nTell me what do u want?: ")
+# handle when user run from file directly
+if __name__ == "__main__":
+    while True:
+        user_input: str = input("\nTell me what do u want?: ")
 
-    # stop process condition
-    if user_input.lower() == "exit":
-        exit()
-    
-    # run algorithm
-    print("user inputted: ", user_input)
-    extraction(user_input)
-    print("Date parameter: ", extract_dates_and_range(user_input))
+        # stop process condition
+        if user_input.lower() == "exit":
+            exit()
+        
+        # perform algorithm
+        init(user_input)